requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.3.2
constantly==15.1.0
cryptography==41.0.4
daphne==4.0.0
Django==4.2.6
django-cors-headers==4.3.0
django-crontab==0.7.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
msgpack==1.0.7
mysqlclient==2.2.0
numpy==1.26.2
packaging==23.2
pandas==2.1.3
pip==23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.2.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
service-identity==23.1.0
setuptools==68.0.0
six==1.16.0
sqlparse==0.4.4
Twisted==23.8.0
# twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
wheel==0.41.2
whitenoise==6.6.0
zope.interface==6.1
